Refactor Docker panels
